requirements.txt
Flask==2.3.0
numpy==1.24.2
pandas==2.0.1
spacy==3.5.0
openai==0.28
nltk==3.8.1
python-docx==0.8.11
plotly==5.15.0
bert-score==0.3.13
scikit-learn==1.2.2
Werkzeug==2.3.2
